optimize

+++ topicModel/plsa/plsa/src/data.h
@@ -12,7 +12,7 @@ class Data {
     size_t GetNumTopics() const { return numTopics_; }
     size_t GetNumDocs() const { return numDocs_; }
     size_t GetNumWords() const { return numWords_; }
-    inline size_t GetAccuDocWords(size_t i, size_t j) const;
+    inline uint8_t GetAccuDocWords(size_t i, size_t j) const;
     inline size_t GetAccuDocs(size_t i) const;
 
   private:
@@ -20,11 +20,11 @@ class Data {
     size_t numDocs_;
     size_t numWords_;
 
-    size_t **accuDocWords_;
+    uint8_t **accuDocWords_;
     size_t *accuDocs_;
 };
 
-size_t Data::GetAccuDocWords(size_t i, size_t j) const {
+uint8_t Data::GetAccuDocWords(size_t i, size_t j) const {
   return accuDocWords_[i][j];
 }
